requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
gradio>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
grad-cam>=1.4.0
tensorboard>=2.14.0
tqdm>=4.66.0
ultralytics>=8.0.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0
h5py>=3.9.0